Introduce a bash completion script
The bash completion script makes it possible to autocomplete commands
and parameters, improving user's experience in the command line.

Currently, the script is only tested/supported on Bash (it may or may
not work on others shells like ksh).

The script is automatically downloaded/sourced when running the setup
script (since the script is dependent on a TorizonCore Builder version,
it is only 'installed' when setting up the environment with the latest
TorizonCore Builder version).

Related-to: TOR-1974

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>