Add extra parameters to torizoncore-builder to identify container
The tcb-env-setup.sh script is now adding some "dynamic" parameters
(i.e. parameters that change on every invocation of tcore-builder) to
allow the tool to identify its own container ID; this is being done to
provide a solution to the bundle command that does not involve using
cgroups information (which are a implementation detail of Docker and
thus might change in the future) to determine the container ID which
is in turn used to gather information about its own execution.

Related-to: TCB-209

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>